+++ protected/humhub/modules/comment/widgets/views/link.php
@@ -2,6 +2,15 @@
 
 use yii\helpers\Html;
 use yii\helpers\Url;
+
+$commentCount = $this->context->getCommentsCount();
+$hasComments = ($commentCount > 0);
+$commentCountSpan = Html::tag('span', ' ('.$commentCount.')', [
+    'class' => 'comment-count',
+    'data-count' => $commentCount,
+    'style' => ($hasComments) ? null : 'display:none'
+]);
+
 ?>
 
 <?php if ($mode == \humhub\modules\comment\widgets\CommentLink::MODE_POPUP): ?>
@@ -10,7 +19,7 @@ use yii\helpers\Url;
         <?= Yii::t('CommentModule.widgets_views_link', "Comment").'('.$this->context->getCommentsCount().')' ?>
     </a>
 <?php elseif(Yii::$app->user->isGuest): ?>
-    <?= Html::a(Yii::t('CommentModule.widgets_views_link', "Comment"), Yii::$app->user->loginUrl, ['data-target' => '#globalModal']) ?>
+    <?= Html::a(Yii::t('CommentModule.widgets_views_link', "Comment").$commentCountSpan, Yii::$app->user->loginUrl, ['data-target' => '#globalModal']) ?>
 <?php else : ?>
-    <?= Html::a(Yii::t('CommentModule.widgets_views_link', "Comment"), "#",['onClick' => "$('#comment_" . $id . "').show();$('#newCommentForm_" . $id . "_contenteditable').focus();return false;"]); ?>
+    <?= Html::a(Yii::t('CommentModule.widgets_views_link', "Comment").$commentCountSpan, "#",['onClick' => "$('#comment_" . $id . "').slideToggle('fast');$('#newCommentForm_" . $id . "').focus();return false;"]); ?>
 <?php endif; ?>