+++ protected/humhub/docs/CHANGELOG.md
@@ -17,6 +17,8 @@ HumHub Change Log
 - Fix: Default user & space module configuration lost after foreign key migration
 - Fix: Respect pinned post when inserting a a new stream entry
 - Fix: Show comments in modal not working
+- Fix: #2381 Comment input not focused on comment link click 
+- Enh: Toggle comment box
 
 
 1.2.0-beta.3 (March 20, 2017)